refactor: extract logic for converting search params to Firebase query conditions into a new file
