+++ src/app/utils/getProtocolHost.ts
@@ -1,0 +1,10 @@
+import { headers } from 'next/headers';
+const getProtocolHost = () => {
+    const headersData = headers();
+    const protocol = headersData.get('x-forwarded-proto');
+    const host = headersData.get('host');
+
+    return { protocol, host };
+}
+
+export default getProtocolHost
